screen-container: rename file
