+++ app/blog/[slug]/page.tsx
@@ -2,7 +2,7 @@ import { notFound } from "next/navigation";
 import { format } from "date-fns";
 import { useTranslations } from "next-intl";
 import { getPostContent, getPostsMetadata } from "@/lib/blog";
-import { ScreenContainer } from "@/components/layout/ScreenContainer";
+import { ScreenContainer } from "@/components/layout/screen-container";
 import { MarkdownPreview } from "@/components/common/markdown-preview";
 import Text from "@/components/texts/text";
