+++ app/event/[id]/exclusive-event-guard.tsx
@@ -22,7 +22,7 @@ import { eventUserRoles } from "@/lib/queries/event-users";
 import { userAddressOptions } from "@/lib/queries/user";
 import { EventPasswordProvider } from "@/components/providers/event-password-provider";
 import { zenaoClient } from "@/app/zenao-client";
-import { ScreenContainer } from "@/components/layout/ScreenContainer";
+import { ScreenContainer } from "@/components/layout/screen-container";
 import { AspectRatio } from "@/components/shadcn/aspect-ratio";
 import { Web3Image } from "@/components/images/web3-image";
