+++ app/not-found.tsx
@@ -2,7 +2,7 @@
 
 import { useTranslations } from "next-intl";
 import Link from "next/link";
-import { ScreenContainerCentered } from "@/components/layout/ScreenContainer";
+import { ScreenContainerCentered } from "@/components/layout/screen-container";
 import { ButtonWithChildren } from "@/components/buttons/ButtonWithChildren";
 import Heading from "@/components/texts/heading";
 import Text from "@/components/texts/text";
