+++ app/settings/page.tsx
@@ -2,7 +2,7 @@ import { auth } from "@clerk/nextjs/server";
 import { dehydrate, HydrationBoundary } from "@tanstack/react-query";
 import { EditUserForm } from "./EditUser";
 import { getQueryClient } from "@/lib/get-query-client";
-import { ScreenContainer } from "@/components/layout/ScreenContainer";
+import { ScreenContainer } from "@/components/layout/screen-container";
 import { userAddressOptions } from "@/lib/queries/user";
 import { profileOptions } from "@/lib/queries/profile";
